life-os/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.17
twilio==9.3.6
google-genai==0.3.0
weaviate-client==4.9.3
neo4j==5.26.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.0
assemblyai==0.34.0
pillow==11.0.0
librosa==0.10.2
soundfile==0.12.1
redis==5.2.0
celery==5.4.0
cachetools==5.5.0
pandas==2.2.3
cryptography==44.0.0
setuptools>=75.0.0 